+++ perl/FSM/Support/LanguageSurfaceContract.pm
@@ -17,6 +17,7 @@ our @EXPORT_OK = qw(
     language_surface_file_surface_entry_keys
     language_surface_file_surfaces_keys
     language_surface_hial_vial_bridge_keys
+    language_surface_vial_execution_keys
     language_surface_nested_presence_key_map
     language_surface_public_top_level_keys
     language_surface_strict_mode_keys
@@ -47,6 +48,7 @@ sub build_language_surface_contract {
         file_surfaces_presence_keys => language_surface_file_surfaces_keys(),
         file_surface_entry_presence_keys => language_surface_file_surface_entry_keys(),
         hial_vial_bridge_presence_keys => language_surface_hial_vial_bridge_keys(),
+        vial_execution_presence_keys => language_surface_vial_execution_keys(),
         default_mode_compatibility_presence_keys => language_surface_default_mode_compatibility_keys(),
         assignments_presence_keys => language_surface_assignments_keys(),
         system_contracts_presence_keys => language_surface_system_contracts_keys(),
@@ -70,6 +72,7 @@ sub language_surface_public_top_level_keys {
             strict_mode
             file_surfaces
             hial_vial_bridge
+            vial_execution
             default_mode_compatibility
             assignments
             system_contracts
@@ -138,6 +141,17 @@ sub language_surface_hial_vial_bridge_keys {
         writes_files
         public_embedding_api
         explicit_nonclaims
+        guidance
+    )];
+}
+
+sub language_surface_vial_execution_keys {
+    return [qw(
+        schema_version status contract_source implementation_entrypoints
+        execution_schema plan_schema replay_schema selected_future_schemas
+        profile capabilities limits
+        fixture writes_files public_embedding_api explicit_nonclaims
+        guidance
     )];
 }
 
@@ -210,6 +224,7 @@ sub language_surface_nested_presence_key_map {
         strict_mode => language_surface_strict_mode_keys(),
         file_surfaces => language_surface_file_surfaces_keys(),
         hial_vial_bridge => language_surface_hial_vial_bridge_keys(),
+        vial_execution => language_surface_vial_execution_keys(),
         default_mode_compatibility => language_surface_default_mode_compatibility_keys(),
         assignments => language_surface_assignments_keys(),
         system_contracts => language_surface_system_contracts_keys(),